pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock_portfolio_analyzer"
version = "0.1.0"
description = "A tool for analyzing stock portfolios"
requires-python = ">=3.9"
dependencies = [
    "pandas",
    "yfinance",
    "pyyaml",
]

[project.scripts]
portfolio-analyzer = "stock_portfolio_analyzer.cli:main"
